+++ Expositum.lua
@@ -4,6 +4,7 @@ local L = EXT.Localization
 local Utils = EXT.Utils
 local Options = EXT.Options
 local Tooltip = EXT.Tooltip
+local AWL = ArcaneWizardLibrary
 
 ----------------------
 --- Local funtions ---
@@ -82,10 +83,10 @@ function expositumFrame:ADDON_LOADED(_, addOnName)
         InitTooltipHooks()
 
         Utils:PrintDebug("Addon fully loaded.")
-		local charKey = GetUnitName("player", true) .. "#" .. GetRealmName()
+		local charKey = AWL.Profiles:GetCharKey()
 
 		if Expositum_Options_v3.profileKeys[charKey]["open-settings"] then
-			Settings.OpenToCategory(EXT.PROFILES_CATEGORY_ID or EXT.MAIN_CATEGORY_ID)
+			Settings.OpenToCategory(EXT.MAIN_CATEGORY_ID)
 
 			Expositum_Options_v3.profileKeys[charKey]["open-settings"] = false
 		end