+++ CHANGELOG.md
@@ -2,5 +2,7 @@
 ### [@project-version@](https://github.com/wow-addon-dev/Expositum/tree/@project-version@) (@build-date@)
 [Full Changelog](@full-changelog@) - [Previous Releases](https://github.com/wow-addon-dev/Expositum/releases)
 
-- Added: Profile settings submenu with account-wide and character profile selection
+Important note: All settings have been reset to their defaults.
+
+- Added: Profile settings with account profile and character profile selection
 - Added: TOC version for patch 12.0.7 [retail]